% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adaptPlogit}
\alias{adaptPlogit}
\title{Adapt "plogit" elements}
\usage{
adaptPlogit(
  M,
  use.rve,
  round.digits = round.digits,
  clubsandwich = FALSE,
  study.var = NULL
)
}
\description{
Adapt "plogit" elements
}
\keyword{internal}
