% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adaptRom}
\alias{adaptRom}
\title{Adapt model results from log ROM to ROM (exp and rename g -> rom)}
\usage{
adaptRom(M, round.digits = 2)
}
\arguments{
\item{M}{List with \code{m} and \code{res}; \code{res} has \code{g}, \code{g.ci}, \code{prediction.ci}.}

\item{round.digits}{Rounding digits for displayed values.}
}
\description{
Adapt model results from log ROM to ROM (exp and rename g -> rom)
}
\keyword{internal}
