% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addAllCombinations}
\alias{addAllCombinations}
\title{Add all trial arm combinations for multiarm trials in NMA.}
\usage{
addAllCombinations(
  data,
  vars.for.id = c("study", "outcome_type", "instrument", "time", "time_weeks", "rating"),
  vars.for.es = c("mean", "sd", "n", "mean_change", "sd_change", "n_change", "event",
    "totaln"),
  condition = "condition",
  condition.specification = "multi",
  groups.column.indicator = c("_arm1", "_arm2")
)
}
\description{
Add all trial arm combinations for multiarm trials in NMA.
}
\keyword{internal}
