% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregateSD}
\alias{aggregateSD}
\title{Aggregate standard deviations of multiple trials}
\usage{
aggregateSD(sd, n, ...)
}
\arguments{
\item{sd}{A numeric vector of standard deviations.}

\item{n}{A numeric vector of sample sizes corresponding to each standard deviation.}

\item{...}{Additional arguments (not used).}
}
\value{
The pooled standard deviation across trials.
}
\description{
This function allows you to aggregate standard deviations of outcome scores
across multiple trials.
}
\details{
Values are aggregated assuming homogeneous variances across studies
(i.e., all trials share the same population standard deviation). The following
formula is used:

\mjtdeqn{s_{\text{pooled}} = \sqrt{ \frac{ \sum (n_i - 1) \cdot s_i^2 }{ \sum (n_i - 1) } }}{s_{\text{pooled}} = \sqrt{ \frac{ \sum (n_i - 1) \cdot s_i^2 }{ \sum (n_i - 1) } }}{s_{\text{pooled}} = \sqrt{ \frac{ \sum (n_i - 1) \cdot s_i^2 }{ \sum (n_i - 1) } }}

where \mjeqn{s_i}{s_i} and \mjeqn{n_i}{n_i} are the standard deviation and
sample size in some study \mjeqn{i}{i}. Please note that aggregation is only
possible when all SD values come from the same scale or instrument.
}
\examples{
\dontrun{
sd <- c(8.77, 9.01, 10.11)
n <- c(120, 15, 80)
aggregateSD(sd, n)
}
}
\keyword{internal}
