% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binaryES}
\alias{binaryES}
\title{Calculate Hedges' g binary outcome data.}
\usage{
binaryES(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{Binary effect size data. Must be \code{Improved_N_ig}, \code{Improved_N_cg}, \code{Rand_N_ig},
\code{Rand_N_cg} (all \code{numeric}).}
}
\description{
Calculate Hedges' g binary outcome data. Only meant to be used as
part of \code{\link{calculateEffectSizes}}.
}
\keyword{internal}
