% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMetaAnalysis.R
\name{blup.runMetaAnalysis}
\alias{blup.runMetaAnalysis}
\title{Best Linear Unbiased Predictions (BLUPs) for 'runMetaAnalysis' models.}
\usage{
\method{blup}{runMetaAnalysis}(x, which = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{runMetaAnalysis}.}

\item{which}{Model for which estimates should be printed. Can be one of \code{"overall"},
\code{"combined"}, \code{"lowest.highest"}, \code{"outliers"},
\code{"influence"}, \code{"threelevel"}, \code{"threelevel.che"}, \code{"ccrem"},
or \code{"ccrem.che"}.}

\item{...}{Additional arguments.}
}
\description{
Generates empirical Bayes (EB) estimates, also known as best linear unbiased
predictions (BLUPs), by merging the fitted values obtained from fixed effects
and estimated contributions of random effects. These estimates represent
the study-specific true effect sizes or outcomes and are accompanied by
standard errors and prediction interval bounds. Uses the \code{\link[metafor:blup]{metafor::blup.rma.uni()}} function
internally.
}
