% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataFormat.R
\name{checkDataFormat}
\alias{checkDataFormat}
\title{Check data format}
\usage{
checkDataFormat(data,
                must.contain = c("study", "condition_arm1",
                                 "condition_arm2", 
                                 "multi_arm1", 
                                 "multi_arm2",
                                 "outcome_type", "instrument",
                                 "time", "time_weeks",
                                 "rating", "mean_arm1", "mean_arm2",
                                 "sd_arm1", "sd_arm2",
                                 "n_arm1", "n_arm2",
                                 "event_arm1", "event_arm2",
                                 "totaln_arm1", "totaln_arm2"),
                variable.class = list("study" = "character", 
                                      "condition_arm1" = "character",
                                      "condition_arm2" = "character", 
                                      "multi_arm1" = "character", 
                                      "multi_arm2" = "character",
                                      "outcome_type" = "character", 
                                      "instrument" = "character",
                                      "time" = "character", 
                                      "time_weeks" = "numeric",
                                      "rating" = "character", 
                                      "mean_arm1" = "numeric", 
                                      "mean_arm2" = "numeric",
                                      "sd_arm1" = "numeric", 
                                      "sd_arm2" = "numeric",
                                      "n_arm1" = "numeric", 
                                      "n_arm2" = "numeric",
                                      "event_arm1" = "numeric", 
                                      "event_arm2" = "numeric",
                                      "totaln_arm1" = "numeric", 
                                      "totaln_arm2" = "numeric"))
}
\arguments{
\item{data}{A \code{data.frame} containint meta-analysis data.}

\item{must.contain}{\code{character} vector, containing all the variable names
the data set should contain. Defaults correspond with the Metapsy data standard.}

\item{variable.class}{\code{list}, defining the required class
for some or all variables. If the class differs in \code{data}, the function will try to
convert the variable to the desired class. Defaults correspond with the
Metapsy data standard.}
}
\value{
\code{checkDataFormat} returns messages that specify if input variables,
values and classes of the variables are as defined.

If default settings are used, \code{checkDataFormat()} can be used in combination
with \code{\link[=checkConflicts]{checkConflicts()}} to determine if a dataset follows the \href{https://docs.metapsy.org/data-preparation/format/}{Metapsy data standard}.
Datasets that are formatted using this standard can be directly used in
the \href{https://tools.metapsy.org/articles/metapsytools#the-analysis-module}{analysis module} in \code{metapsyTools};
for example the \code{\link[=runMetaAnalysis]{runMetaAnalysis()}} function.
}
\description{
This function checks if a \code{data.frame} object conforms with the
\href{https://docs.metapsy.org/data-preparation/format/}{Metapsy data standard}.
}
\details{
The function checks if:
\itemize{
\item{the data set contains all relevant variables and}
\item{variables have the desired class (if not, it tries to convert).}
}
}
\examples{
\dontrun{
data("depressionPsyCtr")

# Example 1: Check with default arguments
checkDataFormat(depressionPsyCtr)

#Example 2: Check for non-default arguments
checkDataFormat(depressionPsyCtr,
                must.contain = c("study", "condition",
                                 "primary", "year"),
                variable.class = list(study = "character",
                                      no.arms = "numeric"))
}


}
\seealso{
\code{\link{expandMultiarmTrials}}
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com},
Pim Cuijpers \email{p.cuijpers@vu.nl}
}
