% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkProblemsNMA}
\alias{checkProblemsNMA}
\title{Check for problems in data sets prepared for netmeta.}
\usage{
checkProblemsNMA(dat.netmeta)
}
\description{
Check for problems in data sets prepared for netmeta.
}
\keyword{internal}
