% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRobRatings.R
\name{checkRobDiscrepancies}
\alias{checkRobDiscrepancies}
\title{Check for inconsistencies between two RoB extraction sheets}
\usage{
checkRobDiscrepancies(data.1, data.2)
}
\arguments{
\item{data.1}{An RoB extraction sheet. Columns of this file must have the same name as in the
\href{https://www.metapsy.org/assets/files/rob-template.xlsx}{RoB extraction sheet template}
provided by the Metapsy initative. If the Metapsy template has been
used, make sure to delete the top rows before importing, so that only the metapsyTools variables
remain as the column names. Required columns are: \code{study}, \code{d1_1}, \code{d1_2}, \code{d1_3}, \code{d1_4}, \code{d1_notes},
\code{d2_5}, \code{d2_6}, \code{d2_7}, \code{d2_8}, \code{d2_9}, \code{d2_notes}, \code{d3_10}, \code{d3_11}, \code{d3_12}, \code{d3_13},
\code{d3_14}, \code{d3_notes}, \code{d4_15}, \code{d4_16}, \code{d4_17}, \code{d4_18}, \code{d4_notes}, \code{d5_19},
\code{d5_20}, \code{d5_21}, \code{d5_22}, \code{d5_23}, \code{d5_24}, \code{d5_notes}.}

\item{data.2}{The same extraction sheet from another (i.e., second) rater.}
}
\value{
If discrepancies are found, the function will return a data frame with
the respective study/studies, along with the diverging ratings (\code{discrepancies} element).
If different studies are included in both sheets, they will be saved under \code{diff.studies}.
}
\description{
Based on prepared extraction sheets by two independent raters, this function allows you
to automatically check for differing studies and/or inconsistencies in the
ratings. Cohen's kappa is calculated automatically, differentiating between
"Yes" vs. "No"/"No Information" ratings to obtain the base rate of agreement.
}
\examples{
\dontrun{
library(readxl)

# Get example extraction sheet from metapsy.org/assets/files/rob_data.xlsx
rob_data <- read_excel("rob_data.xlsx")

# Create second sheet with partly different ratings
rob_data -> rob_data_2
rob_data_2[-1,] -> rob_data_2
rob_data_2[1,"d2_5"] = "NI"
rob_data_2[1,"d4_15"] = "No/PN"

# Check for discrepancies
tmp <- metapsyTools:::checkRobDiscrepancies(rob_data, rob_data_2)
tmp
}

}
\seealso{
\code{\link{createRobRatings}}
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Clara Miguel Sanz \email{clara.miguelsanz@vu.nl},
Pim Cuijpers \email{p.cuijpers@vu.nl}
}
\keyword{internal}
