% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{database2021Subset}
\alias{database2021Subset}
\title{The 'database2021Subset' dataset}
\format{
A \code{data.frame} with 381 rows and 63 variables.
}
\usage{
data("database2021Subset")
}
\description{
An example dataset containing a subset of the 2021 depression trials database.
Its format equals the format of data in the "Metapsy" database.
}
\details{
The dataset has been included to showcase the correct formatting necessary to use \code{metapsyTools}.
}
\author{
Mathias Harrer, Paula Kuper, Pim Cuijpers
}
\keyword{internal}
