% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depressionPsyCtr}
\alias{depressionPsyCtr}
\title{The 'depressionPsyCtr' dataset}
\format{
A \code{data.frame} with 102 rows and 46 variables.
}
\usage{
data("depressionPsyCtr")
}
\description{
An example data set containing a subset of the 2021 depression trials database.
The data set has been formatted following the
\href{https://docs.metapsy.org/data-preparation/format/}{Metapsy data standard}.
}
\details{
The dataset can be used by the \code{runMetaAnalysis} function without any
additional preparation required.
}
\author{
Mathias Harrer, Paula Kuper, Clara Miguel, Pim Cuijpers
}
\keyword{datasets}
