% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Detect}
\alias{Detect}
\title{String detection wrapper}
\usage{
Detect(...)
}
\arguments{
\item{...}{Arguments passed to \code{str_detect}.}
}
\description{
Wrapper for \code{str_detect} in the \code{stringr} package.
}
\keyword{internal}
