% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eb}
\alias{eb}
\title{eb: Empirical Bayes estimates}
\usage{
eb(x, ...)
}
\arguments{
\item{x}{Model}

\item{...}{Other arguments}
}
\description{
Best Linear Unbiased Predictions (BLUPs) for 'runMetaAnalysis' models.
}
\details{
Generates empirical Bayes (EB) estimates, also known as best linear unbiased
predictions (BLUPs), by merging the fitted values obtained from fixed effects
and estimated contributions of random effects. These estimates represent
the study-specific true effect sizes or outcomes and are accompanied by
standard errors and prediction interval bounds.
}
