% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractG}
\alias{extractG}
\title{Extract Hedges \emph{g} values}
\usage{
extractG(x)
}
\arguments{
\item{x}{\code{data.frame} with effect size results.}
}
\description{
Extract Hedges \emph{g} values
}
\keyword{internal}
