% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitCcremModel}
\alias{fitCcremModel}
\title{Fit 'ccrem' model}
\usage{
fitCcremModel(
  data,
  es.var,
  se.var,
  arm.var.1,
  arm.var.2,
  measure.var,
  study.var,
  .raw.bin.es,
  .type.es,
  hakn,
  method.tau.meta,
  method.tau.ci,
  method.tau,
  dots,
  es.binary.raw.vars,
  round.digits,
  nnt.cer,
  which.run,
  mGeneral,
  mCombined,
  use.rve,
  i2.ci.boot,
  nsim.boot
)
}
\description{
Fit 'ccrem' model
}
\keyword{internal}
