% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitCombinedHACEModel}
\alias{fitCombinedHACEModel}
\title{Fit 'combined' complex model}
\usage{
fitCombinedHACEModel(
  which.combine,
  which.combine.var,
  measure.var,
  data,
  study.var,
  multi.study,
  es.var,
  se.var,
  mGeneral,
  .type.es,
  round.digits,
  hakn,
  .raw.bin.es,
  nnt.cer,
  rho.within.study,
  method.tau,
  method.tau.ci,
  dots,
  es.binary.raw.vars,
  arm.var.1,
  arm.var.2,
  phi.within.study,
  n.var.arm1,
  n.var.arm2,
  w1.var,
  w2.var,
  time.var
)
}
\description{
Fit 'combined' complex model
}
\keyword{internal}
