% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitCombinedModel}
\alias{fitCombinedModel}
\title{Fit 'combined' model}
\usage{
fitCombinedModel(
  which.combine,
  which.combine.var,
  data,
  study.var,
  multi.study,
  es.var,
  se.var,
  mGeneral,
  .type.es,
  round.digits,
  hakn,
  .raw.bin.es,
  nnt.cer,
  rho.within.study,
  method.tau,
  method.tau.ci,
  dots,
  es.binary.raw.vars,
  phi.within.study
)
}
\description{
Fit 'combined' model
}
\keyword{internal}
