% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitHighestModel}
\alias{fitHighestModel}
\title{Fit 'highest' model}
\usage{
fitHighestModel(
  data,
  study.var,
  multi.study,
  mGeneral,
  .type.es,
  round.digits,
  .raw.bin.es,
  nnt.cer
)
}
\description{
Fit 'highest' model
}
\keyword{internal}
