% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitInfluenceModel}
\alias{fitInfluenceModel}
\title{Fit 'influence' model}
\usage{
fitInfluenceModel(
  which.influence,
  mComb,
  mGeneral,
  which.run,
  method.tau,
  .raw.bin.es,
  .type.es,
  round.digits,
  nnt.cer,
  rob.data
)
}
\description{
Fit 'influence' model
}
\keyword{internal}
