% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitLowestModel}
\alias{fitLowestModel}
\title{Fit 'lowest' model}
\usage{
fitLowestModel(
  data,
  study.var,
  multi.study,
  mGeneral,
  .type.es,
  round.digits,
  .raw.bin.es,
  nnt.cer,
  rob.data
)
}
\description{
Fit 'lowest' model
}
\keyword{internal}
