% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitOutliersModel}
\alias{fitOutliersModel}
\title{Fit 'outliers' model}
\usage{
fitOutliersModel(
  data,
  study.var,
  multi.study,
  mGeneral,
  .type.es,
  round.digits,
  .raw.bin.es,
  nnt.cer,
  which.run,
  which.outliers,
  method.tau,
  m.for.outliers
)
}
\description{
Fit 'outliers' model
}
\keyword{internal}
