% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitOverallModel}
\alias{fitOverallModel}
\title{Fit 'overall' model}
\usage{
fitOverallModel(
  data,
  es.var,
  se.var,
  arm.var.1,
  arm.var.2,
  measure.var,
  study.var,
  .raw.bin.es,
  .type.es,
  hakn,
  method.tau.meta,
  method.tau.ci,
  method.tau,
  dots,
  es.binary.raw.vars,
  round.digits,
  nnt.cer,
  which.run
)
}
\description{
Fit 'overall' model
}
\keyword{internal}
