% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitRobModel}
\alias{fitRobModel}
\title{Fit 'rob' model}
\usage{
fitRobModel(
  which.run,
  which.rob,
  which.outliers,
  mGeneral,
  mComb,
  low.rob.filter,
  method.tau,
  .raw.bin.es,
  .type.es,
  round.digits,
  nnt.cer
)
}
\description{
Fit 'rob' model
}
\keyword{internal}
