% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitThreeLevelHACEModel}
\alias{fitThreeLevelHACEModel}
\title{Fit 'threelevel' complex model}
\usage{
fitThreeLevelHACEModel(
  data,
  es.var,
  se.var,
  arm.var.1,
  arm.var.2,
  measure.var,
  study.var,
  .raw.bin.es,
  .type.es,
  hakn,
  method.tau.meta,
  method.tau.ci,
  method.tau,
  dots,
  es.binary.raw.vars,
  round.digits,
  nnt.cer,
  which.run,
  mGeneral,
  mCombined,
  use.rve,
  rho.within.study,
  which.combine.var,
  phi.within.study,
  n.var.arm1,
  n.var.arm2,
  w1.var,
  w2.var,
  time.var,
  near.pd,
  i2.ci.threelevel,
  nsim.boot
)
}
\description{
Fit 'threelevel' complex model
}
\keyword{internal}
