% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitWaapWlsModel}
\alias{fitWaapWlsModel}
\title{Fit 'waap-wls' model}
\usage{
fitWaapWlsModel(
  which.run,
  which.waap.wls,
  mGeneral,
  mComb,
  beta.within.study,
  method.tau,
  .raw.bin.es,
  .type.es,
  round.digits,
  nnt.cer,
  rob.data
)
}
\description{
Fit 'waap-wls' model
}
\keyword{internal}
