% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagStudy.R
\name{flagStudy}
\alias{flagStudy}
\title{Flagging engine for internal use}
\usage{
flagStudy(
  smd,
  se,
  rob,
  power = "q1",
  pval = 0.05,
  reference = c("all", "dep", "psy", "ptsd"),
  ...
)
}
\description{
Flagging engine for internal use
}
\keyword{internal}
