% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forestBlup}
\alias{forestBlup}
\title{Function to generate forest plots with EB estimates}
\usage{
forestBlup(
  model,
  which = NULL,
  col.line = "#a7a9ac",
  col.polygon = "#6b58a6",
  leftlab = "Study",
  rightlab = "g [95\% CI]",
  summarylab = "Total (95\% CI)",
  sort = TRUE,
  hetstat = TRUE,
  eb.labels = TRUE
)
}
\description{
Function to generate forest plots with EB estimates
}
\keyword{internal}
