% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forestBlupPlotter}
\alias{forestBlupPlotter}
\title{Internal helper function to generate forest plots with EB estimates}
\usage{
forestBlupPlotter(
  dat,
  res,
  sort,
  col.line,
  col.polygon,
  hetstat,
  leftlab,
  rightlab,
  summarylab,
  M.3l,
  threeLevel,
  eb.labels
)
}
\description{
Internal helper function to generate forest plots with EB estimates
}
\details{
Parts of the code in this function is based on van Aert et al. (2021; Supplement).
}
\references{
van Aert, R. C., Schmid, C. H., Svensson, D., & Jackson, D. (2021).
Study specific prediction intervals for random-effects meta-analysis: A tutorial:
Prediction intervals in meta-analysis. \emph{Research Synthesis Methods, 12}(4), 429-447.
}
\keyword{internal}
