% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{g.binary}
\alias{g.binary}
\title{Calculate Hedges' g using binary outcome data}
\usage{
g.binary(x, cc = TRUE, ...)
}
\arguments{
\item{x}{data}

\item{cc}{Should a continuity correction for zero cells be applied?
Either \code{TRUE} (default; applies the Sweeting et al. (2004) treatment-arm
continuity correction, TACC), \code{FALSE} (no correction), or a single numeric
increment to be added to all four cells (e.g., \code{0.5}).}

\item{...}{Binary effect size data. Data frame must include columns \code{event_arm1}, \code{event_arm2},
\code{totaln_arm1}, \code{totaln_arm2}. See the \href{https://docs.metapsy.org/data-preparation/format/}{Metapsy data standard}.}
}
\description{
Calculates Hedges' g from binary outcome data. Only meant to be used as
part of \code{\link{calculateEffectSizes}}.
}
\keyword{internal}
