% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{g.change.m.sd}
\alias{g.change.m.sd}
\title{Calculate Hedges' g using within-group change data}
\usage{
g.change.m.sd(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{Change score effect size data. Data frame must include columns \code{mean_change_arm1},
\code{mean_change_arm2}, \code{sd_change_arm1}, \code{sd_change_arm2},
\code{n_change_arm1}, \code{n_change_arm2}. See the \href{https://docs.metapsy.org/data-preparation/format/}{Metapsy data standard}.}
}
\description{
Calculate Hedges' g based on change data. Only meant to be used as
part of \code{\link{calculateEffectSizes}}.
}
\keyword{internal}
