% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{g.m.sd}
\alias{g.m.sd}
\title{Calculate Hedges' g using means and standard deviations}
\usage{
g.m.sd(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{Effect size data. Data frame must include columns \code{mean_arm1}, \code{mean_arm2},
\code{sd_arm1}, \code{sd_arm2}, \code{n_arm1}, \code{n_arm2}. See the \href{https://docs.metapsy.org/data-preparation/format/}{Metapsy data standard}.}
}
\description{
Calculate Hedges' g using Mean and Standard Deviation. Only meant to be used as
part of \code{\link{calculateEffectSizes}}.
}
\keyword{internal}
