% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{g.precalc}
\alias{g.precalc}
\title{Forward pre-calculated values of Hedges' g}
\usage{
g.precalc(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{Pre-calculated effect size data. Data frame must include columns \code{precalc_g} and
\code{precalc_g_se}. See the \href{https://docs.metapsy.org/data-preparation/format/}{Metapsy data standard}.}
}
\description{
Forwards pre-calculated values of Hedges' g. Only meant to be used as
part of \code{\link{calculateEffectSizes}}.
}
\keyword{internal}
