% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{i2.gen}
\alias{i2.gen}
\title{Compute I-squared from tau2 using the "generalized" formula
(as used, e.g., in metafor). This formula uses an estimate
of the "typical" within-study variance (v-tilde).}
\usage{
i2.gen(tau2, k, vi)
}
\description{
Compute I-squared from tau2 using the "generalized" formula
(as used, e.g., in metafor). This formula uses an estimate
of the "typical" within-study variance (v-tilde).
}
\keyword{internal}
