% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inpatients}
\alias{inpatients}
\title{The 'inpatients' dataset with 32 clinical trials}
\format{
A \code{data.frame} with 179 rows and 45 variables:
\describe{
  \item{study}{\code{character}, The study label containing the author(s) and year of the study.}
  \item{condition}{\code{character}, The condition of the groups, either "intervention group" or "control group".}
  \item{Cond_spec}{\code{character}, The specific intervention in conditions.}
  \item{is.multiarm}{\code{numeric}, The dichotomized indication if a study has multiple arms or not.}
  \item{no.arms}{\code{numeric}, The number of arms of a study.}
  \item{multiple.arms}{\code{character}, The specification of arms in multiarm studies.}
  \item{Outc_type}{\code{character}, The type of outcome.}
  \item{primary}{\code{numeric}, The indication if a outcome is primary or not.}
  \item{Outc_measure}{\code{character}, The outcome measure used.}
  \item{Time}{\code{character}, The dichotomized time of assessment, either "post" or "FU".}
  \item{Time_weeks}{\code{character}, The assessment time of FU in weeks.}
  \item{year}{\code{numeric}, The year of the study.}
  \item{country}{\code{character}, The country of the study.}
  \item{...}{}
}
}
\usage{
data(inpatients)
}
\description{
An example dataset containing data of 32 clinical trials. Its format equals the format of data in the "metapsy" database that is similar to the long format in R. The dataset also contains columns with study characteristics that are important for effect size calculation and more.
}
\author{
Mathias Harrer, Paula Kuper, Pim Cuijpers
}
\keyword{datasets}
