% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/{metapsyTools}-package.R
\docType{package}
\name{metapsy}
\alias{metapsyTools}
\alias{metapsyTools-package}
\title{metapsyTools: Companion R package for the 'Metapsy' database}
\description{
\code{metapsyTools} is a companion R package for the metapsy database. It aims to faciliate calculation of meta effect sizes for the included trials. For this purpose, data can be checked and converted into the required format to calculate the effect sizes in R. It is then possible to run a meta- analysis over the output effect sizes.
For more information about the metapsy database check out its website: \url{https://evidencebasedpsychotherapies.shinyapps.io/metapsy/}
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com}, Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
\keyword{internal}
