% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metapsyNNT}
\alias{metapsyNNT}
\title{Calculate NNTs}
\usage{
metapsyNNT(d, CER, event.e, n.e, event.c, n.c, names, method)
}
\arguments{
\item{d}{A single numeric or concatenated vector of numerics representing the effect size expressed as
Cohen's \eqn{d} or Hedges' \eqn{g}. If this is the only parameter specified in the function, the method by
Kraemer and Kupfer is used automatically to calculate \eqn{NNT}s.}

\item{CER}{The control group event ratio. Furukawa's method (Furukawa & Leucht, 2011) to calculate \code{NNT}s
from \code{d} requires that the assumed response ("event") ratio in the control group (\eqn{\frac{n_{responders}}{N_{total}}})
is specified. The CER can assume values from 0 to 1. If a value is specified for \code{CER}, Furukawa's method is
used automatically. Argument \code{method} has to be set to \code{"KraemerKupfer"} to override this.}

\item{event.e}{Single number or numeric vector. The number of (favourable) events in the experimental group.}

\item{n.e}{Single number or numeric vector. The number participants in the experimental group.}

\item{event.c}{Single number or numeric vector. The number of (favourable) events in the control group.}

\item{n.c}{Single number or numeric vector. The number of participants in the control group.}

\item{names}{Optional. Character vector of equal length as the vector supplied to \code{d} or \code{event.e} containing
study/effect size labels.}

\item{method}{The method to be used to calculate the NNT from \code{d}. Either \code{"KraemerKupfer"} for the
method proposed by Kraemer and Kupfer (2006) or \code{"Furukawa"} for the Furukawa method (Furukawa & Leucht, 2011).
Please note that the Furukawa's method can only be used when \code{CER} is specified.}
}
\description{
Calculate NNTs (extracted from \code{dmetar})
}
\keyword{internal}
