% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapsyTools-package.R
\docType{package}
\name{metapsyTools}
\alias{metapsyTools}
\alias{metapsyTools-package}
\title{metapsyTools: Companion R package for the 'Metapsy' database}
\description{
\code{metapsyTools} is a companion R package for the \emph{Metapsy} database.
It aims to faciliate the calculation of meta-analytic effect sizes for the
included trials. For this purpose, data can be checked and converted into
the required format to calculate the effect sizes in R. It is then possible
to run a meta-analysis using the output effect sizes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/metapsy-project/metapsyTools}
}

}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com}, Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
\keyword{internal}
