% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaRegression.R
\name{metaRegression}
\alias{metaRegression}
\title{Meta-Regression method for objects of class 'runMetaAnalysis'}
\usage{
metaRegression(x, ...)
}
\arguments{
\item{x}{A model extracted from an object of class \code{runMetaAnalysis}.}

\item{...}{Additional arguments.}
}
\description{
Serves as a wrapper for \code{metareg} or \code{update.rma}, depending
on the class of the fitted model.
}
\examples{
\dontrun{
metaRegression(res$model.combined, ~ rob + scale(year))
}
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
