% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaRegression.R
\name{metaRegression.runMetaAnalysis}
\alias{metaRegression.runMetaAnalysis}
\title{Meta-Regression method for objects of class 'runMetaAnalysis'}
\usage{
\method{metaRegression}{runMetaAnalysis}(x, formula = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{runMetaAnalysis}.}

\item{formula}{not used.}

\item{...}{Additional arguments.}
}
\description{
Prints information that models have to be extracted from \code{runMetaAnalysis}
results objects to run meta-regression.s
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
