% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minimumSNR}
\alias{minimumSNR}
\title{Obtain the minimum signal:noise ratio, given power}
\usage{
minimumSNR(alpha = 0.05, power = 0.8, alternative = "two.sided")
}
\description{
Obtain the minimum signal:noise ratio, given power
}
\keyword{internal}
