% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacementFunctions.R
\name{nnt}
\alias{nnt}
\alias{nnt<-}
\title{Get or set NNT control event rate}
\usage{
nnt(x, ...)

nnt(x) <- value
}
\arguments{
\item{x}{Object of class \code{runMetaAnalysis}.}

\item{...}{Optional arguments (ignored).}

\item{value}{New value for \code{nntCer}.}
}
\value{
For \code{nnt(x)}, the current \code{nntCer} value. For \code{nnt(x) <- value}, the updated object (invisibly).
}
\description{
Generic getter and setter for the NNT control event rate (\code{nntCer})
component of a \code{runMetaAnalysis} object. See \code{\link[metapsyTools]{Replacement functions}}.
}
\keyword{internal}
