% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagStudy.R
\name{plot.flagStudy}
\alias{plot.flagStudy}
\title{Custom plot method for flagStudy objects}
\usage{
\method{plot}{flagStudy}(x, ...)
}
\description{
Custom plot method for flagStudy objects
}
\keyword{internal}
