% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMetaAnalysis.R
\name{plot.runMetaAnalysis}
\alias{plot.runMetaAnalysis}
\title{Plot method for objects of class 'runMetaAnalysis'.}
\usage{
\method{plot}{runMetaAnalysis}(x, which = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{runMetaAnalysis}.}

\item{which}{Model to be plotted. Can be one of \code{"overall"},
\code{"combined"}, \code{"lowest.highest"}, \code{"outliers"},
\code{"influence"}, \code{"baujat"}, \code{"loo-es"} or \code{"loo-i2"}.}

\item{...}{Additional arguments.}
}
\description{
Plot S3 method for objects of class \code{runMetaAnalysis}.
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
