% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateTreatmentCycles.R
\name{plot.simulateTreatmentCycles}
\alias{plot.simulateTreatmentCycles}
\title{Plot method for objects of class 'simulateTreatmentCycles'}
\usage{
\method{plot}{simulateTreatmentCycles}(x, ...)
}
\arguments{
\item{x}{An object of class \code{simulateTreatmentCycles}.}

\item{...}{Additional arguments (not used).}
}
\description{
Plot S3 method for objects of class \code{simulateTreatmentCycles}.
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
