% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupAnalysis.R
\name{plot.subgroupAnalysis}
\alias{plot.subgroupAnalysis}
\title{Plot method for objects of class 'runMetaAnalysis'.}
\usage{
\method{plot}{subgroupAnalysis}(x, which = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{runMetaAnalysis}.}

\item{which}{\code{character}. Subgroup analysis to be plotted (variable name).}

\item{...}{Additional arguments.}
}
\description{
Plot S3 method for objects of class \code{runMetaAnalysis}.
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
