% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConflicts.R
\name{print.checkConflicts}
\alias{print.checkConflicts}
\title{Print method for the 'checkConflicts' function}
\usage{
\method{print}{checkConflicts}(x, ...)
}
\arguments{
\item{x}{A \code{list} object of class \code{checkConflicts}.}

\item{...}{Additional arguments (not used).}
}
\description{
This S3 method prints the studies with potential formatting conflicts.
}
\seealso{
\code{\link{checkConflicts}}
}
\author{
Mathias Harrer & David Daniel Ebert
}
