% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRobRatings.R
\name{print.createRobRatings}
\alias{print.createRobRatings}
\title{Print method for objects of class 'createRobRatings'}
\usage{
\method{print}{createRobRatings}(x, ...)
}
\arguments{
\item{x}{An object of class \code{createRobRatings}.}

\item{...}{Additional arguments.}
}
\description{
Print S3 method for objects of class \code{createRobRatings}.
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
\keyword{internal}
