% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagEffectSizes.R
\name{print.flagEffectSizes}
\alias{print.flagEffectSizes}
\title{Print method for objects of class 'flagEffectSizes'}
\usage{
\method{print}{flagEffectSizes}(x, ...)
}
\arguments{
\item{x}{An object of class \code{flagEffectSizes}.}

\item{...}{Additional arguments.}
}
\description{
Print S3 method for objects of class \code{flagEffectSizes}, as created by \code{\link[=flagEffectSizes]{flagEffectSizes()}}.
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers
\email{p.cuijpers@vu.nl}
}
