% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagStudy.R
\name{print.flagStudy}
\alias{print.flagStudy}
\title{Custom print method for flagStudy objects}
\usage{
\method{print}{flagStudy}(x, ...)
}
\description{
Custom print method for flagStudy objects
}
\keyword{internal}
