% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionMID.R
\name{print.proportionMID}
\alias{print.proportionMID}
\title{Print method for objects of class 'proportionMID'}
\usage{
\method{print}{proportionMID}(x, ...)
}
\arguments{
\item{x}{An object of class \code{proportionMID}.}

\item{...}{Additional arguments.}
}
\description{
Print S3 method for objects of class \code{proportionMID}.
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers
\email{p.cuijpers@vu.nl}
}
