% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateTreatmentCycles.R
\name{print.simulateTreatmentCycles}
\alias{print.simulateTreatmentCycles}
\title{Print method for objects of class 'simulateTreatmentCycles'}
\usage{
\method{print}{simulateTreatmentCycles}(x, ...)
}
\arguments{
\item{x}{An object of class \code{simulateTreatmentCycles}.}

\item{...}{Additional arguments (not used).}
}
\description{
Print S3 method for objects of class \code{simulateTreatmentCycles}.
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com}, Pim Cuijpers
\email{p.cuijpers@vu.nl}
}
