% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupAnalysis.R
\name{print.subgroupAnalysis}
\alias{print.subgroupAnalysis}
\title{Print method for objects of class 'subgroupAnalysis'.}
\usage{
\method{print}{subgroupAnalysis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{subgroupAnalysis}.}

\item{...}{Additional arguments.}
}
\description{
Print S3 method for objects of class \code{subgroupAnalysis}.
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
