% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMetaAnalysis.R
\name{profile.runMetaAnalysis}
\alias{profile.runMetaAnalysis}
\title{Profile Likelihood Plots for 'runMetaAnalysis' models.}
\usage{
\method{profile}{runMetaAnalysis}(fitted, which = NULL, ...)
}
\arguments{
\item{fitted}{An object of class \code{runMetaAnalysis}.}

\item{which}{Model for which estimates should be printed. Can be one of \code{"threelevel"}
or \code{"threelevel.che"}.}

\item{...}{Additional arguments.}
}
\description{
Profiles the restricted log-likelihood of \code{threelevel} and \code{threelevel.che}
models, using the \code{\link[metafor:profile.rma]{metafor::profile.rma()}} function. This functionality
can be used to check if the two heterogeneity variances (\mjeqn{\tau^2}{\tau^2} within
and between studies) were identifiable and correctly estimated.
}
