% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psyCtrSubset}
\alias{psyCtrSubset}
\title{The 'psyCtrSubset' dataset}
\format{
A \code{data.frame} with 503 rows and 60 variables.
}
\usage{
data("psyCtrSubset")
}
\description{
An example dataset containing a subset of the 2021 depression trials database.
Its format equals the format of data in the "metapsy" database, which is similar to the
long format in \emph{R}. The dataset also contains columns with study characteristics that
are important for effect size calculation and more.
}
\details{
The dataset has been included to showcase the correct formatting necessary to use \code{metapsyTools}.
}
\author{
Mathias Harrer, Paula Kuper, Pim Cuijpers
}
\keyword{datasets}
