% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psyCtrSubsetWide}
\alias{psyCtrSubsetWide}
\title{The 'psyCtrSubset' dataset (wide format)}
\format{
A \code{data.frame} with 175 rows and 64 variables.
}
\usage{
data("psyCtrSubsetWide")
}
\description{
An example dataset containing a subset of the 2021 depression trials database.
In constrast to, e.g., the \code{psyCtrSubset} dataset, \code{psyCtrSubsetWide}
is structured in the "wide(r)" format. The dataset also contains columns with study characteristics that
are important for effect size calculation and more.
}
\details{
The dataset has been included to showcase the correct formatting necessary to use \code{metapsyTools}.
}
\author{
Mathias Harrer, Paula Kuper, Clara Miguel, Pim Cuijpers
}
\keyword{internal}
