% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{robSummary}
\alias{robSummary}
\title{Create a RevMan-style risk of bias summary chart}
\usage{
robSummary(data, name.high="High", name.unclear="Unclear",
    name.low="Low", studies, name.missing, table = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} containing a column for each risk of bias criterion, where
rows represent each individual studies. The risk of bias assessment for each criterion in each
study must be coded as a character string. Up to four codes can be used, referring to low risk of bias,
unclear risk of bias, high risk of bias, or missing information. The string used to specify the categories
must be specified in \code{name.high}, \code{name.unclear}, \code{name.low} and/or \code{name.missing},
unless defaults for those parameters are used.}

\item{name.high}{Character specifying how the "high risk of bias" category was coded in \code{data}
(e.g., \code{name.high = "high"}). Default is \code{"High"}.}

\item{name.unclear}{Character specifying how the "unclear risk of bias" category was coded in \code{data}
(e.g., \code{name.unclear = "unclear"}). Default is \code{"Unclear"}.}

\item{name.low}{Character specifying how the "low risk of bias" category was coded in \code{data}
(e.g., \code{name.low = "low"}). Default is \code{"Low"}.}

\item{studies}{A vector of the same length as the number of rows in \code{data} specifying the study
labels for the risk of bias ratings. Only has to be specified when \code{table = TRUE}.}

\item{name.missing}{Character specifying how missing information was coded in \code{data}
(e.g., \code{name.missing} = \code{"missing"}). Default is \code{"Missing"}. All ratings, including missing
information, must be coded as strings, so using \code{NA} in \code{data} to signify missing information
is not valid.}

\item{table}{Should an additional RevMan style risk of bias table be produced? If set to \code{TRUE},
\code{studies} must be specified. \code{FALSE} by default.}
}
\description{
This function generates summary plots for study quality assessments using the
\href{https://bit.ly/2KGQtfG}{Cochrance Risk of Bias Tool}.
Summary plots follow the style of \href{https://bit.ly/30eJK29}{RevMan} Risk of Bias (RoB) summary charts.
}
\details{
The function automatically removes separators like "-" or "." from column names/risk of bias criteria. To produce
a "clean" plot, you may therefore separate words in the column names of the \code{data} data frame using these
symbols (e.g. \code{"Allocation_Concealment"} to return "Allocation Concealment").
}
\references{
Harrer, M., Cuijpers, P., Furukawa, T.A, & Ebert, D. D. (2019).
\emph{Doing Meta-Analysis in R: A Hands-on Guide}. DOI: 10.5281/zenodo.2551803.
\href{https://bookdown.org/MathiasHarrer/Doing_Meta_Analysis_in_R/creating-a-revman-style-risk-of-bias-summary.html}{Chapter 10}
}
\author{
Mathias Harrer & David Daniel Ebert
}
\keyword{internal}
