% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rom.change.m.sd}
\alias{rom.change.m.sd}
\title{Calculate ratio of means (log scale) from change-score means and SDs}
\usage{
rom.change.m.sd(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{columns \code{mean_change_arm1}, \code{mean_change_arm2}, \code{sd_change_arm1}, \code{sd_change_arm2}, \code{n_change_arm1}, \code{n_change_arm2}.}
}
\value{
data.frame with \code{es} (log ROM) and \code{se}.
}
\description{
Only meant to be used as part of \code{\link{calculateEffectSizes}}.
}
\keyword{internal}
