% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rom.m.sd}
\alias{rom.m.sd}
\title{Calculate ratio of means (log scale) using means and standard deviations}
\usage{
rom.m.sd(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{columns \code{mean_arm1}, \code{mean_arm2}, \code{sd_arm1}, \code{sd_arm2}, \code{n_arm1}, \code{n_arm2}.}
}
\value{
data.frame with \code{es} (log ROM) and \code{se}.
}
\description{
Only meant to be used as part of \code{\link{calculateEffectSizes}}.
Returns log(mean_arm1/mean_arm2) and its SE (delta method).
}
\keyword{internal}
