% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rr.precalc}
\alias{rr.precalc}
\title{Forward pre-calculated log-risk ratios}
\usage{
rr.precalc(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{Pre-calculated effect size data. Data frame must include columns
\code{precalc_log_rr} and \code{precalc_log_rr_se}. See the \href{https://docs.metapsy.org/data-preparation/format/}{Metapsy data standard}.}
}
\description{
Forwards pre-calculated log-risk ratios. Only meant to be used as
part of \code{\link{calculateEffectSizes}}.
}
\keyword{internal}
