% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectOutlierModel}
\alias{selectOutlierModel}
\title{Select which model should be used for outliers analysis}
\usage{
selectOutlierModel(which.run, mGeneral, mComb, which.outliers)
}
\description{
Select which model should be used for outliers analysis
}
\keyword{internal}
