% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sendMessage}
\alias{sendMessage}
\title{Send message after fitting model}
\usage{
sendMessage(
  obj,
  model.name = NULL,
  which.run = NULL,
  .type.es = NULL,
  es.type = NULL
)
}
\description{
Send message after fitting model
}
\keyword{internal}
