% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupAnalysis.R
\name{subgroupAnalysis}
\alias{subgroupAnalysis}
\title{Run subgroup analyses}
\usage{
subgroupAnalysis(.model,
                 ...,
                 .which.run = .model$which.run[1],
                 .round.digits = 2,
                 .nnt.cer = NULL,
                 .tau.common = FALSE,
                 .html = TRUE)
}
\arguments{
\item{.model}{An object of class \code{"runMetaAnalysis"}, created by \code{\link{runMetaAnalysis}}.}

\item{...}{<\link[dplyr]{dplyr_data_masking}>. A number of subgroup variables included in the original dataset
provided to \code{\link{runMetaAnalysis}}, separated by commas.}

\item{.which.run}{The model in \code{.model} that should be used for the subgroup analyses. Uses the default
analysis in \code{.model} if no value is specified by the user.}

\item{.round.digits}{\code{numeric}. Number of digits to round the (presented) results by. Default is \code{2}.}

\item{.nnt.cer}{\code{numeric}. Value between 0 and 1, indicating the assumed control group event rate to be used
for calculating NNTs via the Furukawa-Leucht method. If set to \code{NULL} (default),
the value saved in \code{.model} is (re-)used.}

\item{.tau.common}{\code{logical}. Should a common (\code{TRUE}) or subgroup-specific (\code{FALSE}) estimate
of the between-study heterogeneity be calculated when analyzing the subgroups? \code{FALSE} by default. Note that subgroup
analyses based on "multilevel" models automatically assume common heterogeneity estimates.}

\item{.html}{\code{logical}. Should an HTML table be created for the results? Default is \code{TRUE}.}
}
\value{
Returns an object of class \code{"subgroupAnalysis"}. This object includes, among other things,
a \code{data.frame} with the name \code{summary}, in which all subgroup analysis results are summarized.
Other objects are the "raw" subgroup analysis model objects returned.
This allows to conduct further operations on some subgroup analysis specifically.
}
\description{
This function allows to simultaneously conduct different subgroup analyses using
\code{runMetaAnalysis} objects.
}
\details{
For more details see the \href{https://tools.metapsy.org/articles/metapsytools}{Get Started} vignette.
}
\examples{
\dontrun{
data("depressionPsyCtr")

depressionPsyCtr \%>\%
  checkDataFormat() \%>\%
  checkConflicts() \%>\%
  calculateEffectSizes() \%>\% 
  filterPoolingData(condition_arm2 \%in\% 
                      c("wl", "other ctr")) -> data

# Run the meta-analyses
runMetaAnalysis(data) -> res

# Subgroup analysis
subgroupAnalysis(res, condition_arm2, country,
                 .which.run = "combined",
                 .tau.common = TRUE) -> sg
plot(sg, "condition_arm2")
plot(sg, "country")
}

}
\seealso{
\code{\link{runMetaAnalysis}}
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com},
Pim Cuijpers \email{p.cuijpers@vu.nl}
}
