% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMetaAnalysis.R
\name{summary.runMetaAnalysis}
\alias{summary.runMetaAnalysis}
\title{Show details of 'runMetaAnalysis' class objects}
\usage{
\method{summary}{runMetaAnalysis}(object, forest = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{runMetaAnalysis}.}

\item{forest}{\code{logical}. Should a summary forest plot be returned? \code{TRUE} by default.}

\item{...}{Additional arguments passed to \code{\link[meta:forest.meta]{meta::forest()}}.}
}
\description{
S3 method showing analysis settings for objects of class \code{runMetaAnalysis}.
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
