% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{testBaselineImbalance}
\alias{testBaselineImbalance}
\title{Test for baseline imbalances}
\usage{
testBaselineImbalance(
  m_arm1,
  m_arm2,
  sd_arm1,
  sd_arm2,
  n_arm1,
  n_arm2,
  cluster,
  study,
  data = NULL,
  p.adj = "holm",
  ...
)
}
\arguments{
\item{m_arm1}{A vector with the baseline mean(s) in the first arm.}

\item{m_arm2}{A vector with the baseline mean(s) in the second arm.}

\item{sd_arm1}{A vector with the baseline standard deviation(s) in the first arm.}

\item{sd_arm2}{A vector with the baseline standard deviation(s) in the second arm.}

\item{n_arm1}{A vector with the baseline sample size(s) in the first arm.}

\item{n_arm2}{A vector with the baseline sample size(s) in the second arm.}

\item{cluster}{An optional vector of the same length as \code{m_arm1}, \code{m_arm2}, etc.,
encoding to which larger cluster/study a comparison belongs to. If specified,
\emph{p} values of the comparisons will be adjusted according to the method provided in
the \code{p.adj} argument.}

\item{study}{An optional vector with study labels.}

\item{data}{An optional \code{data.frame} that includes the effect size data.}

\item{p.adj}{A character string, specifying the type of \emph{p}-value adjustment. For available
options, see the "Details" section in \code{\link[stats:p.adjust]{stats::p.adjust()}}. Defaults to \code{"holm"}.}

\item{...}{Additional arguments.}
}
\description{
This function returns a SMD for baseline imbalance tests, along with
99\% confidence intervals. If the \code{cluster} argument is specified, \emph{p}-values
can be adjusted using methods available in \code{\link[stats:p.adjust]{stats::p.adjust()}}.
}
\examples{
\dontrun{
# Load example data that follows the Metapsy data standard
data("depressionPsyCtr")

# This is an unexported function, so we need the metapsyTools::: prefix
# Test for differences without p-value adjustment
metapsyTools:::testBaselineImbalance(
  bl_mean_arm1, bl_mean_arm2, 
  bl_sd_arm1, bl_sd_arm2, 
  bl_n_arm1, bl_n_arm2, 
  data = depressionPsyCtr,
  study = study)

# Provide cluster variable: p-values are adjusted within clusters
metapsyTools:::testBaselineImbalance(
  bl_mean_arm1, bl_mean_arm2, 
  bl_sd_arm1, bl_sd_arm2, 
  bl_n_arm1, bl_n_arm2, 
  data = depressionPsyCtr,
  cluster = study,
  study = study)

# Change adjustment method
metapsyTools:::testBaselineImbalance(
  bl_mean_arm1, bl_mean_arm2, 
  bl_sd_arm1, bl_sd_arm2, 
  bl_n_arm1, bl_n_arm2, 
  data = depressionPsyCtr,
  cluster = study,
  study = study, p.adj = "bonferroni")
}

}
\keyword{internal}
