% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{testRandomizedProportion}
\alias{testRandomizedProportion}
\title{Test if the randomized proportion differs from the original allocation ratio}
\usage{
testRandomizedProportion(
  n_arm1,
  n_arm2,
  study,
  data = NULL,
  ratio = "1:1",
  ...
)
}
\arguments{
\item{n_arm1}{A vector in which the sample size of the first trial arm is stored.}

\item{n_arm2}{A vector in which the sample size of the second trial arm is stored.}

\item{study}{An optional vector with study labels.}

\item{data}{An optional \code{data.frame} in which the sample size data is included.}

\item{ratio}{A single string or character vector of the expected allocation ratio.
This ratio should be separated by a colon, e.g. \verb{"1:3}, where the left side
indicates the first trial arm (\code{n_arm1}) and the right side indicates the
second trial arm (\code{n_arm2}). Defaults to \verb{"1:1}.}

\item{...}{Additional arguments.}
}
\description{
This function returns a one-sample proportion test that allows to
examine if the included sample in a study deviated significantly form
the intended allocation ratio.
}
\examples{
\dontrun{
# Load example data that follows the Metapsy data standard
data("depressionPsyCtr")

# This is an unexported function, so we need the metapsyTools::: prefix
# Test if all studies follow the 1:1 allocation ratio
metapsyTools:::testRandomizedProportion(
  n_arm1, n_arm2, data = depressionPsyCtr, study = study) 

# Provide a comparison-specific allocation ratio
ratio <- c("1:1", "3:1", "4:5")
metapsyTools:::testRandomizedProportion(
  n_arm1, n_arm2, data = depressionPsyCtr[1:3,], 
  study = study, ratio = ratio) 
}

}
\keyword{internal}
