% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tryCatch2}
\alias{tryCatch2}
\title{\code{tryCatch} alternative that saves the error message}
\usage{
tryCatch2(expr)

tryCatch2(expr)
}
\arguments{
\item{expr}{R expression.}
}
\description{
\code{tryCatch} alternative that saves the error message

Try function and catch errors
}
\keyword{internal}
