% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateMeta}
\alias{updateMeta}
\title{Import 'update.meta' for internal use}
\usage{
updateMeta(...)
}
\arguments{
\item{...}{Arguments forwarded to \code{update.meta}.}
}
\description{
This function imports the \code{update.meta} function of the \code{meta} package,
which has ceased being an exported function since version 7.0-0.
}
\keyword{internal}
