% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEffectSizes.R
\name{calculateEffectSizes}
\alias{calculateEffectSizes}
\title{Calculate effect sizes}
\usage{
calculateEffectSizes(data,
       comp.id.indicator = "id",
       trt.indicator = "trt",
       funcs = list(meanSD = meanSD,
                    binaryES = binaryES,
                    changeES = changeES),
                    change.sign = NULL)
}
\arguments{
\item{data}{Meta-analysis data set of class \code{expandMultiarmTrials}, created using \code{expandMultiarmTrials}.}

\item{comp.id.indicator}{\code{character}, column name of the variable storing the comparison ID; typically created by \code{expandMultiarmTrials}.}

\item{trt.indicator}{\code{character}, column name of the variable storing the treatment indicator (treatment 1 or 2); typically created by \code{expandMultiarmTrials}.}

\item{funcs}{\code{list} of functions. These functions will be used to calculate the effect sizes (Hedges' \emph{g}) based on the raw data (see Details).}

\item{change.sign}{\code{character}. Name of a \code{logical} column in \code{data}, encoding if the
sign of a calculated effect size should be reversed (\code{TRUE}) or not (\code{FALSE}). Set to \code{NULL} (default) if
no changes should be made.}
}
\value{
\code{calculateEffectSizes} returns the meta-analysis data set as class \code{data.frame} in wide format (if results are saved to a variable). It also generates the following columns, wich are added to the data:
\itemize{
\item{\code{es} calculated effect sizes for each comparison.}
\item{\code{se} calculated standard error of the effect size.}
\item{\code{study.id} a study-specific ID variable.}
\item{\code{study} a study-specific variable containing its name. For multiarm studies, this variable specifies the active treatment arm used to calculate the effect.}
}
}
\description{
This is a function to calculate effect sizes of meta-analysis data created by \code{expandMultiarmTrials}.
}
\details{
By default, the small-sample bias corrected standardized mean difference (Hedges' \emph{g}) is calculated from:
\itemize{
 \item {the mean, SD and N}
 \item {binary outcome data (i.e. response counts) and}
 \item {change scores}
}
Other functions can be added to the list provided to \code{funcs}. However, results of the function must result in a \code{data.frame}
with the same number of rows as in \code{data}, and two columns: one for the calculated \emph{g} value (named \code{es}) and its standard error (named \code{se}).
In rows for which no fitting raw data was supplied, the resulting \code{data.frame} should contain \code{NA}.

For more details see the help vignette: \code{vignette("metapsyTools")}
}
\examples{
\dontrun{
#Example 1: use function in default mode; requires data created by expandMultiarmTrials
data("inpatients")
inpatients \%>\%
    expandMultiarmTrials() \%>\%
    calculateEffectSizes()

#Example 2: further use to pool effect sizes
library(meta)
inpatients \%>\%
  checkDataFormat() \%>\%
  expandMultiarmTrials() \%>\%
  calculateEffectSizes() \%>\%
  filterPoolingData(primary==1) \%>\%
  metagen(es, se, studlab=study, comb.fixed=FALSE, data=.)


# Example 3: use for 3-level model
library(metafor)
library(dplyr)
inpatients \%>\%
  checkDataFormat ()\%>\%
  expandMultiarmTrials() \%>\%
  calculateEffectSizes() \%>\%
  dplyr::mutate(es.id = 1:nrow(.)) \%>\%
  metafor::rma.mv(es, se^2, data = ., slab = study.id,
                  random = ~ 1 | study.id/es.id, test = "t")
}


}
\seealso{
\code{\link{expandMultiarmTrials}}
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com}, Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
