% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataFormat.R
\name{checkDataFormat}
\alias{checkDataFormat}
\title{Check data format}
\usage{
checkDataFormat(data,
           must.contain = c("study", "condition",
                            "Cond_spec", "is.multiarm",
                            "no.arms", "multiple_arms",
                            "Outc_measure", "Time", "primary",
                            "Time_weeks", "sr_clinician"),
           variable.contains = list("condition" = c("ig", "cg"),
                                    "is.multiarm" = c(0,1)),
           variable.class = list("study" = "character",
                                 "condition" = "character",
                                 "Cond_spec" = "character",
                                 "is.multiarm" = "numeric",
                                 "no.arms" = "numeric",
                                 "multiple_arms" = "character",
                                 "primary" = "numeric",
                                 "Time" = "character",
                                 "Time_weeks" = "character",
                                 "sr_clinician" = "character",
                                 "Post_M" = "numeric",
                                 "Post_SD" = "numeric",
                                 "Post_N" = "numeric",
                                 "Rand_N" = "numeric",
                                 "Improved_N" = "numeric",
                                 "Change_m" = "numeric",
                                 "Change_SD" = "numeric",
                                 "Change_N" = "numeric"),
           groups.column.indicator = c("_trt1", "_trt2"),
           data.format = NULL)
}
\arguments{
\item{data}{Meta-analysis data stored as a data.frame.}

\item{must.contain}{\code{character} vector, containing all the variable names the data set should contain.}

\item{variable.contains}{\code{list}, defining which values should only be contained in a variable.}

\item{variable.class}{\code{list}, defining the required data class for some or all variables. If the class differs, the function will try to convert the variable to the desired class.}

\item{groups.column.indicator}{\code{character}. If the dataset is in wide format: a character vector with two elements, representing the suffix used to
differentiate between the first and second treatment in a comparison.}

\item{data.format}{\code{character}. Either \code{"long"} or \code{"wide"}, depending on the format of the dataset in
\code{data}. \code{NULL} by default, which lets the user define the format after the function has been called.}
}
\value{
\code{checkDataFormat} returns messages that specify if input variables, values and classes of the variables are as defined.
}
\description{
This is a function to check the format of meta-analysis data with class \code{data.frame} for its further applicability to the \code{expandMultiarmTrials} function.
}
\details{
The function ckecks if:
 \itemize{
 \item{the data set contains all relevant variables}
 \item{variables contain desired values only and}
 \item{variables are of desired type (if not, it tries to convert them).}
 }
 If the dataset format is wide, please specify all variable/column names
 \emph{without} the group indicator suffix (e.g. for \code{"Cond_spec_trt1"} and
 \code{"Cond_spec_trt2"}, simply use \code{"Cond_spec"}).
}
\examples{
\dontrun{
data("inpatients")

# Example 1: Check data with default arguments
checkDataFormat(inpatients)

#Example 2: Check for specified variables and corresponding values and classes
checkDataFormat(inpatients,
                must.contain = c("study", "condition",
                                 "primary", "year"),
                variable.contains = list("condition" = c("ig","cg")),
                variable.class = list(study = "character",
                                      no.arms = "numeric"))

#Example 3: Convert variable class as predefined
checkDataFormat(inpatients,
                must.contain = c("primary", "study"),
                variable.class = list(primary = "integer"))

#Example 4: Wide format
data("psyCtrSubsetWide")
checkDataFormat(psyCtrSubsetWide)
}


}
\seealso{
\code{\link{expandMultiarmTrials}}
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com}, Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
