% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isNAorNaN}
\alias{isNAorNaN}
\title{Tests if value is \code{NA} or \code{NaN}}
\usage{
isNAorNaN(x)
}
\arguments{
\item{expr}{R expression.}
}
\description{
Tests if value is \code{NA} or \code{NaN}
}
\keyword{internal}
