% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meanSD}
\alias{meanSD}
\title{Calculate Hedges' g using Mean and Standard Deviation.}
\usage{
meanSD(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{Effect size data. Must be \code{Post_M_trt1}, \code{Post_M_trt2}, \code{Post_SD_trt1},
\code{Post_SD_trt2}, \code{Post_N_trt1}, \code{Post_N_trt2} (all \code{numeric}).}
}
\description{
Calculate Hedges' g using Mean and Standard Deviation. Only meant to be used as
part of \code{\link{calculateEffectSizes}}.
}
\keyword{internal}
