% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiarmExpander}
\alias{multiarmExpander}
\title{Expander function for multiarm trials}
\usage{
multiarmExpander(
  study,
  condition.specification,
  group.indicator,
  group.names,
  id
)
}
\arguments{
\item{study}{data of one study, for one unique assessment point.}

\item{condition.specification}{trial condition specification.}

\item{group.indicator}{group indicator (IG or CG).}
}
\description{
Expands multiarm trial.
}
\keyword{internal}
