% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacementFunctions.R
\name{Replacement functions}
\alias{Replacement functions}
\alias{data<-}
\alias{data}
\alias{which.run}
\alias{near.pd}
\alias{es.measure}
\alias{es.type}
\alias{es.var}
\alias{se.var}
\alias{es.binary.raw.vars}
\alias{method.tau}
\alias{hakn}
\alias{study.var}
\alias{arm.var.1}
\alias{arm.var.2}
\alias{measure.var}
\alias{low.rob.filter}
\alias{method.tau.ci}
\alias{round.digits}
\alias{which.combine}
\alias{which.combine.var}
\alias{which.outliers}
\alias{which.influence}
\alias{which.rob}
\alias{nnt.cer}
\alias{rho.within.study}
\alias{phi.within.study}
\alias{w1.var}
\alias{w2.var}
\alias{time.var}
\alias{vcov}
\alias{use.rve}
\alias{html}
\alias{lower.is.better}
\alias{selmodel.steps}
\alias{rerun}
\alias{which.run<-}
\alias{es.measure<-}
\alias{es.type<-}
\alias{es.var<-}
\alias{se.var<-}
\alias{es.binary.raw.vars<-}
\alias{method.tau<-}
\alias{hakn<-}
\alias{study.var<-}
\alias{arm.var.1<-}
\alias{arm.var.2<-}
\alias{measure.var<-}
\alias{low.rob.filter<-}
\alias{method.tau.ci<-}
\alias{round.digits<-}
\alias{which.combine<-}
\alias{which.combine.var<-}
\alias{which.outliers<-}
\alias{which.influence<-}
\alias{which.rob<-}
\alias{nnt.cer<-}
\alias{rho.within.study<-}
\alias{phi.within.study<-}
\alias{w1.var<-}
\alias{w2.var<-}
\alias{time.var<-}
\alias{vcov<-}
\alias{use.rve<-}
\alias{html<-}
\alias{lower.is.better<-}
\alias{selmodel.steps<-}
\alias{near.pd<-}
\title{Replacement functions for "runMetaAnalysis" results objects}
\usage{
data(x) <- value
which.run(x) <- value
es.measure(x) <- value
es.type(x) <- value
es.var(x) <- value
se.var(x) <- value
es.binary.raw.vars(x) <- value
method.tau(x) <- value
hakn(x) <- value
study.var(x) <- value
arm.var.1(x) <- value
arm.var.2(x) <- value
measure.var(x) <- value
low.rob.filter(x) <- value
method.tau.ci(x) <- value
which.combine(x) <- value
which.combine.var(x) <- value
which.outliers(x) <- value
which.influence(x) <- value
which.rob(x) <- value
nnt.cer(x) <- value
rho.within.study(x) <- value
phi.within.study(x) <- value
w1.var(x) <- value
w2.var(x) <- value
vcov(x) <- value
near.pd(x) <- value
use.rve(x) <- value
html(x) <- value
lower.is.better(x) <- value
selmodel.steps(x) <- value
rerun(m)
}
\arguments{
\item{x}{Object of class \code{runMetaAnalysis}.}

\item{value}{Value of one of the arguments in \code{runMetaAnalysis} or
\code{correctPublicationBias}}

\item{m}{(Adapted) object of class \code{runMetaAnalysis}.}
}
\description{
Once a model has been fitted using \code{runMetaAnalysis},
replacement functions are defined for each function argument. This
allows to quickly tweak one or more analysis settings, which are implemented
once the \code{rerun} function is called.
}
\examples{
\dontrun{
data("depressionPsyCtr")
depressionPsyCtr \%>\%
  checkDataFormat() \%>\%
  checkConflicts() \%>\%
  calculateEffectSizes() \%>\% 
  filterPoolingData(condition_arm2 \%in\% 
                      c("wl", "other ctr")) -> data

m <- runMetaAnalysis(data, "combined")

# Compare results when other tau^2 estimator is used
method.tau(m) <- "DL"
rerun(m)
}
}
\seealso{
\code{\link{runMetaAnalysis}}, \code{\link{correctPublicationBias}}
}
