% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rr.binary}
\alias{rr.binary}
\title{Calculate the log-risk ratio using binary outcome data}
\usage{
rr.binary(x, cc = 0.5, ...)
}
\arguments{
\item{x}{data}

\item{cc}{Should a continuity correction for zero cells be applied?
Either \code{FALSE} or the increment to be added. Default is 0.5.}

\item{...}{Binary effect size data. Data frame must include columns \code{event_arm1}, \code{event_arm2},
\code{totaln_arm1}, \code{totaln_arm2}. See the \href{https://docs.metapsy.org/data-preparation/format/}{Metapsy data standard}.}
}
\description{
Calculate the log risk ratio using binary outcome data. Only meant to be used as
part of \code{\link{calculateEffectSizes}}.
}
\keyword{internal}
