% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMetaAnalysis.R
\name{runMetaAnalysis}
\alias{runMetaAnalysis}
\title{Run different types of meta-analyses}
\usage{
runMetaAnalysis(data,

                # Models to run
                which.run = c("overall", "combined",
                              "lowest.highest", "outliers",
                              "influence", "rob", "threelevel",
                              "threelevel.che"),
                              
                # Effect size measure
                es.measure = c("g", "RR"),
                es.type = c("precalculated", "raw"),
                es.var = ifelse(identical(es.measure[1], "g"), 
                                ".g", ".log_rr"),
                se.var = ifelse(identical(es.measure[1], "g"), 
                                ".g_se", ".log_rr_se"),
                es.binary.raw.vars = 
                  c(".event_arm1", ".event_arm2",
                    ".totaln_arm1", ".totaln_arm2"),
                    
                # Estimator of the heterogeneity variance
                method.tau = "REML",
                method.tau.ci = "Q-Profile",
                i2.ci.boot = FALSE,
                nsim.boot = 5e3,
                hakn = TRUE,
                
                # Data specifications
                study.var = "study",
                arm.var.1 = "condition_arm1",
                arm.var.2 = "condition_arm2",
                measure.var = "instrument",
                low.rob.filter = "rob > 2",
                round.digits = 2,
                rob.data = NULL,
                
                # Model specifications
                which.combine = c("arms", "studies"),
                which.combine.var = "multi_arm1",
                which.outliers = c("overall", "combined"),
                which.influence = c("overall", "combined"),
                which.rob = c("overall", "combined"),
                nnt.cer = 0.2,
                rho.within.study = 0.6,
                phi.within.study = 0.9,
                w1.var = ifelse(identical(es.measure[1], "g"), 
                                "n_arm1", "totaln_arm1"),
                w2.var = ifelse(identical(es.measure[1], "g"), 
                                "n_arm2", "totaln_arm2"),
                time.var = "time_weeks",
                vcov = c("simple", "complex"),
                near.pd = FALSE,
                use.rve = TRUE,
                
                # Output
                html = TRUE,
                
                # Additional arguments
                ...)
}
\arguments{
\item{data}{\code{data.frame}. Effect size data in the wide format, as
created by \code{\link{calculateEffectSizes}}.}

\item{which.run}{\code{character}. Selection of models to be calculated.
See 'Details'.}

\item{es.measure}{\code{character}. Should meta-analyses be calculated using the
bias-corrected standardized mean difference (\code{"g"}; default), or using
risk ratios (\code{"RR"})? Meta-analyses will only be conducted using comparisons
that contain non \code{NA} values in the \code{es.var} and \code{se.var} columns.}

\item{es.type}{\code{character}. Should pre-calculated or raw event data (i.e.
the Mantel-Haenszel method) be used for meta-analyses of risk ratios?
Can be set to \code{"precalculated"} (default) or \code{"raw"}.}

\item{es.var}{\code{character}. Specifies the name of the variable containing the (pre-
calculated) effect size data in \code{data}. When \code{es.measure = "g"}, this is set to
\code{.g} by default; \code{".log_rr"} is used when \code{es.measure = "RR"}. The default settings
correspond with the standard output of \code{\link[=calculateEffectSizes]{calculateEffectSizes()}}.}

\item{se.var}{\code{character}. Specifies the name of the variable containing the (pre-calculated)
standard errors (square root of the variance) of the effect size metric defined
in \code{es.var}. If \code{es.measure = "g"}, this is automatically set to \code{.g_se}; if
\code{es.measure = "RR"}, \code{".log_rr_se"} is used. The default settings
correspond with the standard output of \code{\link[=calculateEffectSizes]{calculateEffectSizes()}}.}

\item{es.binary.raw.vars}{\code{character} vector, defining the column names in
which the (1) raw event counts in the experimental group, (2) raw event counts in the
control/reference group, (3) sample size in the experimental group, and (4) sample size
of the control/reference group are stored.
Defaults correspond with the standard output of \code{\link[=calculateEffectSizes]{calculateEffectSizes()}}.}

\item{method.tau}{\code{character}. A character string indicating
which method is used to estimate the between-study variance (tau-squared)
and its square root (tau). Either \code{"REML"} (default), \code{"DL"},
\code{"PM"}, \code{"ML"}, \code{"HS"}, \code{"SJ"}, \code{"HE"}, or \code{"EB"}, can be abbreviated (see
\code{\link[meta]{metagen}}). Use \code{"FE"} to use a fixed-effect/"common effect" model.}

\item{method.tau.ci}{\code{character}. A character string indicating which method is used to estimate the
confidence interval of the between-study heterogeneity variance \mjeqn{\tau^2}{\tau^2}. Either \code{"Q-Profile"} (default and recommended; \href{http://www.wvbauer.com/lib/exe/fetch.php/articles:viechtbauer2007b.pdf}{Viechtbauer, 2017}),
\code{"BJ"}, \code{"J"}, or \code{"PL"} can be abbreviated. See \code{\link[meta]{metagen}} and \code{\link[metafor]{rma.uni}} for details.}

\item{i2.ci.boot}{\code{logical}. Confidence intervals for \mjeqn{\tau^2}{\tau^2} as calculated by the Q-Profile method are not
directly applicable for three-level models, in which two heterogeneity variance components are estimated. By default,
this argument is therefore set to \code{FALSE}, and no confidence intervals around \mjeqn{\tau^2}{\tau^2} and \mjeqn{I^2}{I^2} are provided for the
\code{"threelevel"} and \code{"threelevel.che"} model.
If this argument is set to \code{TRUE}, parametric bootstrapping is used to calculate confidence intervals around the between- and
within-study heterogeneity estimates (\mjeqn{\tau}{\tau} and \mjeqn{I^2}{I^2}). Please note that this can take several minutes,
depending on the number of effect sizes. If \code{\link[=correctPublicationBias]{correctPublicationBias()}} is used and \code{i2.ci.boot} is \code{TRUE},
bootstrapping will also be used to calculate confidence intervals around the \mjeqn{G^2}{G^2} statistic
(\href{https://academic.oup.com/biostatistics/article/12/1/122/391113}{Rücker et al., 2011})
used in the limit meta-analysis (note that \mjeqn{G^2}{G^2} is printed as \mjeqn{I^2}{I^2} in this package).}

\item{nsim.boot}{\code{numeric} Number of bootstrap samples to be drawn when \code{i2.ci.boot} is \code{TRUE}. Defaults to
5000.}

\item{hakn}{\code{logical}. Should the Knapp-Hartung adjustment for effect size significance tests be used? Default is \code{TRUE}.}

\item{study.var}{\code{character}. The name of the variable in \code{data} in which the study IDs are stored.}

\item{arm.var.1}{\code{character}. The name of the variable in \code{data} in which the condition (e.g. "guided iCBT")
of the \emph{first} arm within a comparison are stored.}

\item{arm.var.2}{\code{character}. The name of the variable in \code{data} in which the condition (e.g. "wlc")
of the \emph{second} arm within a comparison are stored.}

\item{measure.var}{\code{character}. The name of the variable in \code{data}
in which the instrument used for the comparison is stored.}

\item{low.rob.filter}{\code{character}. A filtering statement by which to
include studies for the "low RoB only" analysis. Please note that the name of
the variable must be included as a column in \code{data}.}

\item{round.digits}{\code{numeric}. Number of digits to round the (presented) results by. Default is \code{2}.}

\item{rob.data}{\code{list}. Optional list detailing how risk of bias data should be appended to the model (see Details).}

\item{which.combine}{\code{character}. Should multiple effect sizes within one study be pooled
on an \code{"arms"} (default) or \code{"studies"} level? When a study is a multi-arm trial, setting
\code{which.combine = "arms"} will aggregate the effect sizes for each trial arm individually before pooling;
the \code{which.combine.var} argument can be used to control which effects within a study should be aggregated.
When \code{which.combine = "studies"}, one overall aggregated effect is created for each study. This setting is preferable
from a statistical perspective, since it ensures that all pooled effects can be assumed to be independent.}

\item{which.combine.var}{\code{character}. Additional grouping variable within studies to be used for the \code{"combined"}
analysis when \code{which.combine = "arms"}. If the specified variable differs within one study (as defined
by \code{study.var}), effects will be aggregated separately for each unique value in \code{which.combine.var}.
Defaults to \code{"multi_arm1"}, the variable which encodes multi-arm intervention conditions in the
\href{https://docs.metapsy.org/data-preparation/format/#standard-variables}{Metapsy data standard}.}

\item{which.outliers}{\code{character}. Which model should be used to conduct outlier analyses? Must be
\code{"overall"} or \code{"combined"}, with \code{"overall"} being the default.}

\item{which.influence}{\code{character}. Which model should be used to conduct influence analyses? Must be
\code{"overall"} or \code{"combined"}, with \code{"overall"} being the default.}

\item{which.rob}{\code{character}. Which model should be used to conduct the "low risk of bias only" analyses? Must be
\code{"overall"} or \code{"combined"}, with \code{"overall"} being the default.}

\item{nnt.cer}{\code{numeric}. Value between 0 and 1, indicating the assumed control group event rate to be used
for calculating NNTs via the Furukawa-Leucht method.}

\item{rho.within.study}{\code{numeric}. Value between 0 and 1, indicating the assumed correlation of effect sizes
within studies. This is relevant to combine effect sizes for the \code{"combined"} analysis type, and used to estimate
the variance-covariance matrices needed for the conditional and hierarchical effects three-level model. Default is \code{0.6}.}

\item{phi.within.study}{\code{numeric}. Value between 0 and 1, indicating the assumed one-week autocorrelation of effect sizes.
This is only used when \code{vcov="complex"} to approximate the variance-covariance matrices needed for the \code{"combined"} and
\code{"threelevel.che"} model. Default is 0.9. See "Details".}

\item{w1.var}{\code{character}. Name of the variable in \code{data} in which the sample
sizes of the first arm are stored. See "Details".}

\item{w2.var}{\code{character}. Name of the variable in \code{data} in which
the sample sizes of the second arm are stored. See "Details".}

\item{time.var}{\code{character}. Name of the variable in \code{data} in which the assessment time point is stored.
Should be expressed as weeks since randomization; other units (e.g. days, months) are also possible if \code{phi.within.study}
is specified accordingly. See "Details".}

\item{vcov}{\code{character}. For the \code{"combined"} and \code{"threelevel.che"} model, should variance-covariance matrices
(representing the dependency structure of the data) be approximated using a heterogeneous compound symmetry (\code{"simple"}; default)
or unstructured matrix structure (\code{"complex"})? See "Details".}

\item{near.pd}{\code{logical}. If at least one of the study variance-covariance matrices constructed
when \code{vcov="complex"} is not positive definite/invertible, should the \code{\link[Matrix]{nearPD}} function
be used to compute the nearest positive definite matrix? Default is \code{FALSE}.}

\item{use.rve}{\code{logical}. Should robust variance estimation be used to calculate confidence intervals and tests of three-level models?
\code{TRUE} by default.}

\item{html}{\code{logical}. Should an HTML table be created for the results? Default is \code{TRUE}.}

\item{...}{Additional arguments.}
}
\value{
Returns an object of class \code{"runMetaAnalysis"}. This object includes, among other things,
a \code{data.frame} with the name \code{summary}, in which all results are summarized - including the
studies which were removed for some analysis steps. Other objects are the "raw" model objects returned
by all selected analysis types. This allows to conduct further operations on some models specifically (e.g.
run a meta-regression by plugging one of the model objects in \code{meta:::update.meta}.
}
\description{
This wrapper function allows to simultaneously pool effect sizes using
different meta-analytic approaches.
\loadmathjax
}
\details{
The \code{runMetaAnalysis} function is a wrapper for several types of meta-analytic models
that are typically used. It allows to run all of these models in one step in order to generate results
that are somewhat closer to being "publication-ready".

By default, the following models are calculated:

\itemize{
\item \code{"overall"}. Runs a generic inverse-variance (random-effects) model. All included
effect sizes are treated as independent. When \code{es.measure = "RR"} and \code{es.type = "raw"},
the Mantel-Haenszel method is used for pooling instead.
\item \code{"combined"}. Pools all effect sizes within one study (defined by \code{study.var}) before
pooling. This ensures that all effect sizes are independent (i.e., unit-of-analysis error &
double-counting is avoided). To combine the effects, one has to assume a correlation of effect sizes
within studies, empirical estimates of which are typically not available.
\item \code{"lowest.highest"}. Runs a meta-analysis, but with only (i) the lowest and (ii) highest
effect size within each study included.
\item \code{"outlier"}. Runs a meta-analysis without statistical outliers (i.e. effect sizes for which
the confidence interval does not overlap with the confidence intervall of the overall effect).
\item \code{"influence"}. Runs a meta-analysis without influential cases (see \code{\link[metafor]{influence.rma.uni}} for
details).
\item \code{"rob"}. Runs a meta-analysis with only low-RoB studies included.
\item \code{"threelevel"}. Runs a multilevel (three-level) meta-analysis model, with effect sizes nested
in studies.
\item \code{"threelevel.che"}. Runs a multilevel (three-level) meta-analysis model, with effect sizes nested
in studies. Variance-covariance matrices of each study with two or more effect sizes are estimated using
\code{rho.within.study} as the assumed overall within-study correlation. This imputation allows to run a "correlated and
hierarchical effects" (CHE) model, which is typically a good approximation for data sets with unknown and/or
complex dependence structures.
}

Internally, the \code{overall}, \code{combined}, \code{lowest.highest}, \code{outlier}, \code{influence} and \code{rob}
models are fitted by calling the \code{\link[meta:metagen]{meta::metagen()}} or \code{\link[meta:metabin]{meta::metabin()}} function,
respectively, in \strong{\{meta\}} (Balduzzi, Rücker & Schwarzer, \href{https://pubmed.ncbi.nlm.nih.gov/31563865/}{2019}).
The \code{threelevel} and \code{threelevel.che} models are implemented using \code{\link[metafor:rma.mv]{metafor::rma.mv()}}
in \strong{\{metafor\}} (Viechtbauer, \href{https://www.jstatsoft.org/article/view/v036i03}{2005}).

Outlier selection is implemented using the \code{\link[dmetar:find.outliers]{dmetar::find.outliers()}} function,
and influence analyses using the \code{\link[dmetar:InfluenceAnalysis]{dmetar::InfluenceAnalysis()}} function. The latter function is a wrapper for
\code{\link[metafor:influence.rma.uni]{metafor::influence.rma.uni()}}.

\mjeqn{~}{~}
\subsection{Simple or complex variance-covariance approximation}{

The \code{vcov} argument controls if the effect size dependencies within the data
should be approximated using a \code{"simple"} (default) or more \code{"complex"} (but potentially more accurate)
method. This argument is only relevant for the \code{"combined"} and \code{"threelevel.che"} models. The default
"simple" method constructs variance-covariance matrices \mjeqn{\Sigma_k}{\Sigma_k} for each study using a
constant sampling correlation \mjeqn{\rho}{\rho} (defined by \code{rho.within.study}), which is identical across all studies, outcomes, and time points.
This simplifying assumption is part of the formulation of the CHE model originally provided by
Pustejovsky and Tipton (\href{https://link.springer.com/article/10.1007/s11121-021-01246-3}{2022}).

Naturally, employing a common value of \mjeqn{\rho}{\rho} across all studies may not be reasonable
in some analyses, and other information may be available to better approximate the effect size dependencies
in the collected data. Setting \code{vcov} to \code{"complex"} allows to assume that correlations between effect
sizes may differ conditional on the type of dependency. This means that the variance-covariance matrix \mjeqn{\Sigma_k}{\Sigma_k} of some study \mjeqn{k}{k} is
approximated by an unstructured matrix with varying \mjeqn{\rho_{ij}}{\rho_{ij}} (instead of a
heterogeneous compound symmetry matrix with fixed \mjeqn{\rho}{\rho}, as is used when \code{vcov="simple"}).

\mjtdeqn{\small \begin{array}{ccc} \texttt{vcov="simple"} & \texttt{vcov="complex"} & \\\ \Sigma_k =\left[ \begin{array}{cccc} \sigma^2_1 & & & \\\ \rho \sigma_2 \sigma_1 & \sigma^2_2 & & \\\ \rho \sigma_3 \sigma_1 & \rho \sigma_3 \sigma_2 & \sigma^2_3 & \\\ \rho \sigma_4 \sigma_1 & \rho \sigma_4 \sigma_2 & \rho \sigma_4 \sigma_3 & \sigma^2_4 \end{array} \right] & \left[ \begin{array}{cccc} \sigma^2_1 & & & \\\ \rho_{21} \sigma_2 \sigma_1 & \sigma^2_2 & & \\\ \rho_{31} \sigma_3 \sigma_1 & \rho_{32} \sigma_3 \sigma_2 & \sigma_3 & \\\ \rho_{41} \sigma_4 \sigma_1 & \rho_{42} \sigma_4 \sigma_2 & \rho_{43} \sigma_4 \sigma_3 & \sigma^2_4 \end{array} \right] &  \end{array}}{\begin{array}{ccc}\texttt{vcov="simple"} & \texttt{vcov="complex"} & \\\\\ \Sigma_k = \begin{bmatrix} \sigma^2_1 \\\\\ \rho \sigma_2 \sigma_1 & \sigma^2_2 & & \\\\\ \rho \sigma_3 \sigma_1 & \rho \sigma_3 \sigma_2 & \sigma^2_3 & \\\\\ \rho \sigma_4 \sigma_1 & \rho \sigma_4 \sigma_2 & \rho \sigma_4 \sigma_3 & \sigma^2_4 \end{bmatrix} & \Sigma_k = \begin{bmatrix} \sigma^2_1 & & & \\\\\ \rho_{21} \sigma_2 \sigma_1 & \sigma^2_2 & & \\\\\ \rho_{31} \sigma_3 \sigma_1 & \rho_{32} \sigma_3 \sigma_2 & \sigma^2_3 & \\\\\ \rho_{41} \sigma_4 \sigma_1 & \rho_{42} \sigma_4 \sigma_2 & \rho_{43} \sigma_4 \sigma_3 & \sigma^2_4 \end{bmatrix} &  \end{array}}{}

For example, setting \code{vcov = "complex"} allows to additionally incorporate assumed correlations specific to multiple testing over time
(e.g. correlations between effects at post-test and long-term follow-up). The value provided in
\code{phi.within.study} represents the (auto-)correlation coefficient \mjeqn{\phi}{\phi}, which serves
as a rough estimate of the re-test correlation after 1 week. When a vector of follow-up lengths
is provided in \code{time.var}, this allows to model the gradual decrease in correlation between measurements
over time. Furthermore, it is possible to calculate a correlation coefficient \mjeqn{\rho_w}{\rho_w} for
multi-arm trials, which is directly proportional to the size of each individual trial arm. When all trial
arms have the same size, meaning that each arm's weight \mjeqn{w}{w} is identical,
\mjeqn{\rho_w}{\rho_w} is known to be 0.5. Multiarm weights \mjeqn{w}{w} (and thus
\mjeqn{\rho_w}{w}) can be derived if the \code{w1.var} and \code{w2.var} variables,
containing the sample size of each study arm, are provided.

Using the complex approximation method increases the risk that at least one studies' \mjeqn{\Sigma_k}{\Sigma_k} matrix
is not positive definite. In this case, the function automatically switches back to the
constant sampling correlation approximation.

\mjeqn{~}{~}
}

\subsection{Replacement functions}{

Once a model has been fitted using \code{runMetaAnalysis}, \strong{replacement functions}
are defined for each function argument. This allows to quickly tweak one or
more analysis settings, which are implemented once the \code{rerun} function is called.
Say that we saved the results of \code{runMetaAnalysis} in an object \code{m}. If, for example, we want
to check the results using a different estimator of \mjeqn{\tau^2}{\tau^2}, leaving
all other settings the same, we could run e.g. \code{method.tau(m) <- "PM"}, followed by
\code{rerun(m)}. This would provide results using the Paule-Mandel estimator. A list of all available
setting replacement functions is provided \href{https://tools.metapsy.org/reference/replacement-functions}{here}.

\mjeqn{~}{~}
}

\subsection{Risk of Bias data}{

Using the \code{rob.data} argument, it is possible to specify which variables in the data set provided in
\code{data} contain Risk of Bias (ROB) assessment information. If specified, this allows to generate
forest plots with added ROB information when using \code{\link[=plot.runMetaAnalysis]{plot.runMetaAnalysis()}}. ROB information added this
way can also be used to create ROB summary plots using \code{\link[=createRobSummary]{createRobSummary()}}.

The object provided to \code{rob.data} must be a \code{list} element with these elements:
\itemize{
\item \code{domains}: A vector of characters, specifying variables that contain ratings for different ROB domains (e.g. allocation concealment, missing data handling, ...).
\item \code{domain.names} (\emph{Optional}): A vector of characters with the same length as \code{domains}, which provides long-format labels for each included domain.
\item \code{overall.rob} (\emph{Optional}): A single character specifying the variable in \code{data} that contains the overall ROB rating.
\item \code{categories}: A vector of characters, specifying how ROB ratings were coded in the selected variables (e.g., \verb{"low}, \code{"high"}, \code{"unclear"}).
\item \code{symbols} (\emph{Optional}): A vector of single-letter characters (or symbols) that should be used when plotting ROB rating in the forest plot.
\item \code{colors} (\emph{Optional}): A vector of characters of the same length as \code{categories}, specifying colors for each rating code.
}

For concrete usage examples with this functionality, see "Examples".

For more details see the \href{https://tools.metapsy.org/articles/metapsytools}{Get Started} vignette.
}
}
\examples{
\dontrun{
data("depressionPsyCtr")
library(meta)

depressionPsyCtr \%>\%
  checkDataFormat() \%>\%
  checkConflicts() \%>\%
  calculateEffectSizes() \%>\% 
  filterPoolingData(condition_arm2 \%in\% 
                      c("wl", "other ctr")) -> data

# Run the meta-analyses
runMetaAnalysis(data) -> res

# Use replacement function to show results for
# differing settings
method.tau(res) <- "PM"
hakn(res) <- FALSE
rerun(res)

# Show summary
res

# Show forest plot (by default, "overall" is used)
plot(res)

# Show forest plot of specific analysis
plot(res, "outliers")
plot(res, "threelevel")
plot(res, "baujat")
plot(res, "influence")
plot(res, "lowest.highest")

# Extract specific model and do further calculations
# (e.g. meta-regression on 'year')
metaRegression(res$model.overall, ~ scale(year))

# Conduct a subgroup analysis
subgroupAnalysis(res, country)

# Correct for publication bias/small-study effects
correctPublicationBias(res)

# For the combined analysis, set which.combine to
# "studies" here, so that all effects in a study are aggregated
# first before pooling
data \%>\% 
  runMetaAnalysis(which.combine = "studies") \%>\% 
  plot("combined")
  
# Define ROB data to be added to the models
robData = list(
  # Names of ROB variables included in 'data'
  domains = c("sg", "ac", "ba", "itt"),
  # Long-format labels for each ROB domain
  domain.names = c("Sequence Generation", 
                   "Allocation Concealment", 
                   "Blinding of Assessors", 
                   "ITT Analyses"),
  # Codes used to rate the risk of bias (sr=self-report)
  categories = c("0", "1", "sr"),
  # Symbols that should be used for these codes in forest plots
  symbols = c("-", "+", "s"),
  # Colors to be used in forest plots for each of these codes
  colors = c("red", "green", "yellow"))

# Re-run model with appended ROB data
res = runMetaAnalysis(data, rob.data = robData) 

# Generate forest plot with ROB data
plot(res, "combined")

# Create a summary plot
createRobSummary(res, 
                 name.low = "1", 
                 name.high = "0", 
                 name.unclear = "sr",
                 which.run = "combined")

# Run meta-analysis using raw response rate data
data \%>\% 
  runMetaAnalysis(es.measure = "RR",
                  es.type = "raw")
}

}
\seealso{
\code{\link{calculateEffectSizes}},
\code{\link{subgroupAnalysis}}. \code{\link{correctPublicationBias}},
\code{\link{metaRegression}}
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
